+++ src/features/featureB/components/card/index.ts
@@ -1,0 +1,6 @@
+import Card from './Card';
+import CardDetail from './CardDetail';
+import CardDetailNavigating from './CardDetailNavigating';
+import CardListWrapper from './CardListWrapper';
+
+export { Card, CardDetail, CardDetailNavigating, CardListWrapper };
