+++ src/features/featureB/components/modal/index.ts
@@ -1,0 +1,5 @@
+import CompletedModal from './CompletedModal';
+import GuideModal from './GuideModal';
+import VerificationModal from './VerificationModal';
+
+export { CompletedModal, GuideModal, VerificationModal };
