+++ src/features/featureB/types/index.ts
@@ -1,0 +1,1 @@
+export type Phase = 'initial' | 'eventInfo' | 'navigate';
